# lib/api-zod/src/generated/types/academySessionItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Palace Protocol Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface AcademySessionItem {
  id: number;
  title: string;
  date: string;
  time: string;
  location: string;
  description: string;
  capacity: number;
}