# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Palace Protocol Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 1 */
  volume: number;
  /** @minimum 0 */
  price: number;
  description: string;
  /** @nullable */
  coverImage?: string | null;
  coverPending?: boolean;
}