# lib/api-zod/src/generated/types/enrollmentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Palace Protocol Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface EnrollmentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  message?: string;
}