# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Palace Protocol Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface MemberUpdate {
  name?: string;
  email?: string;
  phone?: string;
  status?: string;
}