# lib/api-zod/src/generated/types/siteContentUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Palace Protocol Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteContentUpdate {
  heroHeadline?: string;
  heroSubheadline?: string;
  aboutHeading?: string;
  aboutBody1?: string;
  aboutBody2?: string;
  /** @nullable */
  ceoPhoto?: string | null;
  terms?: string;
  policies?: string;
  pillar1Numeral?: string;
  pillar1Title?: string;
  pillar1Description?: string;
  pillar2Numeral?: string;
  pillar2Title?: string;
  pillar2Description?: string;
  pillar3Numeral?: string;
  pillar3Title?: string;
  pillar3Description?: string;
}