# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Palace Protocol Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  whatsappNumber: string;
  email: string;
  location: string;
  sessionTimes: string;
  ceoName: string;
  ceoTitle: string;
}